Fix escaping issues
We need to account for the fact that the attribute names are first
interpreted by the shell and then interpreted by Nix itself when
building. The latter means we have to enclose every attribute in path
in double quotes (""), the former means we need to quote the entire
argument in single quotes to prevent interpretation by shell. We
_hope_ that the attribute paths don't contain double quotation marks,
since AFAIU there's no way to make nix evaluate or build attributes
which do have them.